Code entropy and generalize moran.

+++ model/src/main/scala/fr/iscpif/spacematters/model/package.scala
@@ -33,9 +33,8 @@ package object model {
   }
 
   type State = Matrix[Cell]
-
   type Position = (Int, Int)
-  type Quantity[T] = T => Int
+  type Quantity[T] = T => Double
 
   implicit def colorToCellQuantity(c: Color): Quantity[Cell] = c.cellColor.get(_)
   implicit def matrixToSeqOfSeq[T](m: Matrix[T]) = m.matrix
